Improve alerts summary performance (#4376)
* Reduced alerts summary bucket size

* Added Alerts Summary table class

* Added Alerts Summary table class

* Added base configuration tables

* feat: add structures for aggregation queries of summary tables of modules without selected agent

* Split extended information and added agents tables

* Code cleaning

* Added optional chaining to conditionals

* Added missing alerts summary definitions

* Delete hidden vis and fix merge conflicts

* Fix merge

* Add changelog

(cherry picked from commit 395fb8a485e8bcac6f96a4fbc8dd07100f7dc2c7)
