+++ server/lib/reporting/summary-tables-definitions/index.ts
@@ -1,0 +1,6 @@
+import agentsSummary from './agents';
+import overviewSummary from './overview';
+export default {
+    agentsSummary,
+    overviewSummary
+}
